+++ index.html
@@ -12,7 +12,7 @@
 <body>
     <div id="root">Not Rendered</div>
 
-    <script type="module" src="App.js"></script>
+    <script type="module" src="src/App.js"></script>
 
 </body>
