+++ src/config.js
@@ -1,0 +1,137 @@
+const resturantList = [
+    {
+        type: "resturant",
+        data: {
+            id: "restaurant-1",
+            name: "Restaurant 1",
+            address: "613 Main St, Los Angeles, NY 54788",
+            phone: "664-321-6065",
+            cuisine: "Indian",
+            rating: 2,
+            price_range: "$$$",
+            image: "https://source.unsplash.com/7/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            id: "restaurant-2",
+            name: "Restaurant 2",
+            address: "560 Main St, Phoenix, NY 96473",
+            phone: "839-776-8780",
+            cuisine: "Indian",
+            rating: 2,
+            price_range: "$$",
+            image: "https://source.unsplash.com/random/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            id: "restaurant-3",
+            name: "Restaurant 3",
+            address: "152 Main St, Los Angeles, AZ 69945",
+            phone: "487-125-3876",
+            cuisine: "Indian",
+            rating: 3,
+            price_range: "$",
+            image: "https://source.unsplash.com/u/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            id: "restaurant-4",
+            name: "Restaurant 4",
+            address: "838 Main St, Chicago, TX 45575",
+            phone: "148-407-8317",
+            cuisine: "Mexican",
+            rating: 1,
+            price_range: "$$$$",
+            image: "https://source.unsplash.com/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            "id": "restaurant-5",
+            "name": "Restaurant 5",
+            "address": "143 Main St, Chicago, TX 94536",
+            "phone": "864-920-1928",
+            "cuisine": "Italian",
+            "rating": 4,
+            "price_range": "$",
+            "image": "https://source.unsplash.com/o/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            "id": "restaurant-6",
+            "name": "Restaurant 6",
+            "address": "133 Main St, Los Angeles, CA 10067",
+            "phone": "573-790-8458",
+            "cuisine": "Italian",
+            "rating": 1,
+            "price_range": "$$$",
+            "image": "https://source.unsplash.com/random/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            "id": "restaurant-7",
+            "name": "Restaurant 7",
+            "address": "966 Main St, Chicago, IL 96456",
+            "phone": "482-194-8408",
+            "cuisine": "Chinese",
+            "rating": 3,
+            "price_range": "$",
+            "image": "https://source.unsplash.com/t/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            "id": "restaurant-8",
+            "name": "Restaurant 8",
+            "address": "441 Main St, Houston, IL 57169",
+            "phone": "180-118-3732",
+            "cuisine": "American",
+            "rating": 4,
+            "price_range": "$$$$",
+            "image": "https://source.unsplash.com/l/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            "id": "restaurant-9",
+            "name": "Restaurant 9",
+            "address": "793 Main St, Los Angeles, IL 69273",
+            "phone": "196-233-3084",
+            "cuisine": "Chinese",
+            "rating": 2,
+            "price_range": "$$$$",
+            "image": "https://source.unsplash.com/r/800x600"
+        }
+    },
+    {
+        type: "resturant",
+        data: {
+            "id": "restaurant-10",
+            "name": "Restaurant 10",
+            "address": "195 Main St, Los Angeles, NY 11957",
+            "phone": "756-317-2492",
+            "cuisine": "Indian",
+            "rating": 5,
+            "price_range": "$$$",
+            "image": "https://source.unsplash.com/800x600"
+        }
+    }
+];
+
+
+
+
+export default resturantList
